+++ src/main/java/org/example/flow_analyzer/models/Transaction.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -17,7 +18,7 @@ public class Transaction {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
-    private double operation; //Сумма
+    private BigDecimal operation; //Сумма
     private LocalDate localDate;
     private String description;//Описание
     @ManyToOne
